# src/configs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InsecureRestAPI",
    "version": "1.0",
    "description": "<b>InsecureRestAPI - An insecure NodeJS/Express/MongoDB REST API for educational purposes.</b>\n\nYou can select a development or production server to test the API. Most operations require authentication via a user specific JWT token. To retrieve a JWT token for a user you can use the Site <code>'/api/v1/site/sign-in'</code> operation below and then copy the value of the <b>accessToken</b> field. This value can then be entered when you click on the <b>Authorize</b> button or lock icons.\n\n[View Schema](/docs/openapi.json)",
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "contact": {
      "name": "Kevin A. Lee (kadraman)",
      "url": "https://github.com/kadraman"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "InsecureRestAPI - development "
    }
  ],
  "paths": {
...
    "/api/v1/site/backup-newsletter-db": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Backup the newsletter database",
        "description": "Compress and backup the newsletter database to the specified file",
        "operationId": "backupNewsletterDb",
        "parameters": [
          {
            "name": "filePath",
            "in": "query",
            "description": "The file to backup the database to. Cannot be empty.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "success"
        }
      },
      "failure": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failure"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "failure"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access Token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}